+++ packages/node/src/indexer/fetch.service.ts
@@ -12,6 +12,7 @@ import {
   ApiService,
   getModulos,
   getLogger,
+  Header,
 } from '@subql/node-core';
 import { EthereumBlock, SubqlDatasource } from '@subql/types-ethereum';
 import { SubqueryProject } from '../configure/SubqueryProject';
@@ -29,15 +30,12 @@ const BLOCK_TIME_VARIANCE = 5000;
 
 const INTERVAL_PERCENT = 0.9;
 
-const logger = getLogger(`EthFetchService`);
-
 @Injectable()
 export class FetchService extends BaseFetchService<
   SubqlDatasource,
   IEthereumBlockDispatcher,
   EthereumBlock
 > {
-  private lastFinalizedHeight?: number;
   constructor(
     private apiService: ApiService,
     nodeConfig: NodeConfig,
@@ -46,7 +44,7 @@ export class FetchService extends BaseFetchService<
     @Inject('IBlockDispatcher')
     blockDispatcher: IEthereumBlockDispatcher,
     dictionaryService: EthDictionaryService,
-    private unfinalizedBlocksService: UnfinalizedBlocksService,
+    unfinalizedBlocksService: UnfinalizedBlocksService,
     eventEmitter: EventEmitter2,
     schedulerRegistry: SchedulerRegistry,
   ) {
@@ -58,6 +56,7 @@ export class FetchService extends BaseFetchService<
       dictionaryService,
       eventEmitter,
       schedulerRegistry,
+      unfinalizedBlocksService,
     );
   }
 
@@ -65,25 +64,9 @@ export class FetchService extends BaseFetchService<
     return this.apiService.unsafeApi;
   }
 
-  protected async getFinalizedHeight(): Promise<number> {
+  protected async getFinalizedHeader(): Promise<Header> {
     const block = await this.api.getFinalizedBlock();
-
-    const header = ethereumBlockToHeader(block);
-    const newFinalizedBlockHeight = header.blockHeight;
-    logger.debug(
-      `Rpc finalized height, ${newFinalizedBlockHeight} | current lastFinalizedHeight, ${this.lastFinalizedHeight}`,
-    );
-    // Rpc could return finalized height below last finalized height due to unmatched nodes
-    // See how this could happen in https://gist.github.com/jiqiang90/ea640b07d298bca7cbeed4aee50776de
-    // We want to ensure indexer not stall due to this
-    if (
-      this.lastFinalizedHeight === undefined ||
-      newFinalizedBlockHeight > this.lastFinalizedHeight
-    ) {
-      this.lastFinalizedHeight = newFinalizedBlockHeight;
-      this.unfinalizedBlocksService.registerFinalizedBlock(header);
-    }
-    return this.lastFinalizedHeight;
+    return ethereumBlockToHeader(block);
   }
 
   protected async getBestHeight(): Promise<number> {