+++ packages/node/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-- Improve indexer could stall due to rpc finalized height could be smaller than previous result
+- Update with `node-core`, change `getFinalizedHeight` to `getFinalizedHeader`. Fix indexer could stall due to rpc finalized height could be smaller than previous result. (#2487)
 
 ## [4.7.1] - 2024-07-10
 ### Fixed